Minor improvements to the ODBC program
The result processing now uses a separate class for formatting the
result instead of doing it explicitly. This allows the same code to be
used by both the normal SQL queries as well as the data streaming.

Fixed bugs in the SQL and made it detect more types (e.g AUTO_INCREMENT
columns). Added a GenericTranslator class for doing imports based on
only the ODBC catalog function data.