Fix PostgreSQL import SQL
The auto-increment field was always defined as a primary key even if it
wasn't the sole primary key. This means that indexes on the table must
be defined when the table is being created. With this change, no
redundant temporary indexes end up being created.

Added support for importing geometry types that can be converted into
WKT form. The line, line segment, box and circle types are imported as
TEXT types since they are not convertible to a WKT representation.